requirements.txt
Django==1.9.7
django-debug-toolbar==1.4
django-mptt==0.8.4
PyMySQL==0.7.4
sqlparse==0.1.19
